% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_label}
\alias{add_label}
\title{Replace acronyms by labels for selected columns}
\format{
a data frame
}
\usage{
add_label(x, vars)
}
\arguments{
\item{x}{a data frame obtained using \code{stbl}}

\item{vars}{a character vector indicating the columns for which
short names should be replaced by complete names}
}
\description{
Replace acronyms by labels for selected columns
}
