% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_eq}
\alias{check_eq}
\alias{verif_eq}
\title{Check the equilibrium of the SAM}
\usage{
verif_eq(x, diff = c("abs", "rel"), n = NULL)
}
\arguments{
\item{x}{an object of class \code{sam}}

\item{diff}{the difference either absolute (\code{"abs"}) or relative
(\code{"rel"})}

\item{n}{the number of lines to print}
}
\value{
a data frame with a line for each account, lines and
columns' margins and the relative and absolute differences
}
\description{
The margins of lines (ressources) and columns (expenses) of the SAM
are computed and their equality are checked
}
