% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{condsummary}
\alias{condsummary}
\title{Summarise using a function and a set of factors}
\format{
a series or a data frame
}
\usage{
condsummary(x, gps, FUN, dim = c("levels", "table"), ...)
}
\arguments{
\item{x}{a data frame obtained using \code{stbl}}

\item{gps}{a list of factors}

\item{FUN}{the function}

\item{dim}{with the default (\code{"levels"}) returns a data frame with
all the combinations of the factors and a \code{valeur} column
containing the result, if \code{"table"} returns a vector that can
be added to the original data frame}

\item{...}{further arguments passed to \code{FUN} (for example \code{na.rm})}
}
\description{
The \code{valeur} column of the table is transformed according to a
function, for all the combinations of a set of factors
}
