% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ipc}
\alias{ipc}
\title{Consumer Price Index}
\usage{
ipc(x, base = 2023, deflate = TRUE)
}
\arguments{
\item{x}{a numerical vector containing years or a data frame}

\item{base}{the base, ie the year for which the index is equal to 1}

\item{deflate}{if \code{TRUE} numerical values of the data frame are
divided by the ipc, otherwise, the ipc is added to the data
frame}
}
\value{
a numerical vector containing the ipc for the requested
years or the \code{x} data frame augmented by a \code{ipc} column
}
\description{
Extract the cpi for requested years and add a \code{ipc} column in a
table returned by \code{syntbl} or deflate the \code{valeur} column by the
\code{ipc}
}
\examples{
ipc(2010:2015)
stbl(2020:2022, tableau = "produits",
     variable = c("cf", "fbcf"), naf = "A1") |>
     ipc(base = 2020)
stbl(2020:2022, tableau = "produits",
     variable = c("cf", "fbcf"), naf = "A1") |>
     ipc(base = 2020, deflate = FALSE)
}
