% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pop}
\alias{pop}
\title{Population}
\usage{
pop(x, deflate = TRUE)
}
\arguments{
\item{x}{a numerical vector containing years or a data frame}

\item{deflate}{if \code{TRUE} numerical values of the data frame are
divided by the population, otherwise, the population is added
to the data frame}
}
\value{
a numerical vector containing the population for the
requested years or the \code{x} data frame augmented by a \code{pop}
column
}
\description{
Extract the population (in millions inhabitants) for requested
years or add a pop column in a table returned by \code{syntbl}
}
\examples{
pop(2010:2015)
stbl(2020:2022, tableau = "produits",
     variable = c("cf", "fbcf"), naf = "A1") |>
     pop()
stbl(2020:2022, tableau = "produits",
     variable = c("cf", "fbcf"), naf = "A1") |>
     pop(deflate = FALSE)
}
