% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/sam.R
\name{summary.sam}
\alias{summary.sam}
\alias{as.matrix.sam}
\alias{sam}
\title{Social Account Matrix}
\usage{
\method{summary}{sam}(object, ...)

\method{as.matrix}{sam}(x, ...)

sam(
  an = NULL,
  naf = c("A17", "A38", "A1", "A4", "A10", "Sections"),
  tot_gen = TRUE,
  isbl = FALSE,
  sep_ent = FALSE,
  net_acc = TRUE,
  net_tax = TRUE,
  net_cap = TRUE,
  rev_int = FALSE,
  simpl_dist = FALSE,
  trace = FALSE,
  unit = c("G", "M", "K", "PIB")
)
}
\arguments{
\item{object, x}{an object of class \code{sam},}

\item{...}{further arguments for the \code{summary} and the \code{as.matrix}
method (currently unused)}

\item{an}{a year}

\item{naf, isbl, net_acc, sep_ent, net_tax, simpl_dist, net_cap}{see
\code{stbl}}

\item{tot_gen}{a boolean, if \code{TRUE}, a general total is added for
each account}

\item{rev_int}{a boolean, if \code{TRUE}, intermediary incomes (primary
income and disposable income are added to the SAM)}

\item{trace}{a boolean, if \code{TRUE} (the default is \code{FALSE}), some
information messages are printed while computing the SAM}

\item{unit}{monetary unit, \code{K} for thousands of euros, \code{M} for
millions of euros, \code{G} for billions of euros (the default) and
\code{PIB} for a 100 base for the PIB,}
}
\value{
an object of class \code{mcs} which inherits the \code{data.frame}
class
}
\description{
Generate a SAM for France, many arguments enable the user to
customize the result
}
\details{
\code{sam} returns a data frame in long format (one line for each
value), with a \code{"sam"} class. The \code{as.matrix} method transforms the
SAM in a large format, each row and column being an account.
}
