% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sam2tt}
\alias{sam2tt}
\title{SAM formated as a markdown table}
\usage{
sam2tt(x, height = 1, theme = TRUE)
}
\arguments{
\item{x}{a \code{sam} object}

\item{height}{the height of the lines}

\item{theme}{a boolean, if \code{TRUE} the \code{theme_sam} function is
applied}
}
\description{
The tinytable package is used and functions of this package can be
used to customize the result. The table can then be coerced in
several format, especially LaTeX and html
}
