% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setaname}
\alias{setaname}
\title{Set the name of one variable and return the data frame}
\format{
a data frame
}
\usage{
setaname(x, old, new)
}
\arguments{
\item{x}{a data frame}

\item{old}{the old name, a character}

\item{new}{the new name, a character}
}
\description{
Set the name of one variable and return the data frame
}
