% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stbl.R
\name{stbl}
\alias{stbl}
\alias{stbl2sam}
\title{Synthetic Table}
\format{
a \code{stbl} object, which inherits from data frame
containing :
\itemize{
\item for the long format: each line contains a variable
\item for the large format:
\itemize{
\item for production sectors and products, each line contains a
combination of years and sectors / products and each column is
a variable
\item for the TEE: each line is a variable, each column is a
combination of an agent and either ressource or expense
}
}
}
\usage{
stbl(
  an = 2023,
  tableau = c("produits", "branches", "tei", "matprod", "tee"),
  unit = c("G", "M", "K", "E"),
  format = c("long", "marges", "large"),
  sgn = NULL,
  label = NULL,
  variables = NULL,
  net_tax = TRUE,
  isbl = FALSE,
  naf = NULL,
  net_acc = TRUE,
  sep_ent = FALSE,
  net_cap = TRUE,
  simpl_dist = FALSE,
  agents = NULL,
  comptes = NULL
)

stbl2sam(x)
}
\arguments{
\item{an}{une annee}

\item{tableau}{the table}

\item{unit}{monetary unit, \code{K} for thousands of euros, \code{M} for
millions of euros, \code{G} for billions of euros (the default)}

\item{format}{\code{large}for a table with one line for a product /
sector for the TES, one line for a variable for the TEE. \code{long}
for a table with one line for each value. \code{large} for a table
which returns the production and the intermediate consumptions
for sectors or by products}

\item{sgn}{\code{pos} to select variables that are ressources and \code{neg}
to select variables that are expenses and \code{diff} (only for the
TEE) for the difference between ressources and expenses}

\item{label}{a character vector that indicates labels that should
be printed in the table}

\item{variables}{the variables to be extracted (by default, all the
variables are returned)}

\item{net_tax}{a boolean, if \code{TRUE}, the default, taxes on
production and products are computed as the difference between
taxes and subventions}

\item{isbl}{a boolean, if \code{FALSE}, the default, ISBL are merged
with households,}

\item{naf}{the activity nomenclature}

\item{net_acc}{a boolean, if \code{TRUE}, the default, investment and
stock variations are merged}

\item{sep_ent}{a boolean, if \code{TRUE}, financial and non financial
societies are separated,}

\item{net_cap}{a boolean, if \code{TRUE}, capital income is splited in
EBE, en revenu mixte brut and properties income}

\item{simpl_dist}{a boolean, if \code{TRUE} (the default is \code{FALSE}),
the distribution operations are simplified. All taxes and
social cotisations are received by the State, all social
prestations are paid by the State and received by households,
income tax is only paid by enterprises and households}

\item{agents}{a subset of economic agents for the TEE}

\item{comptes}{a subset of accounts for the TEE}

\item{x}{an object of class \code{stbl} for the \code{stbl2sam} function}
}
\description{
Extract part of the series of the two synthetic tables, the TEE
(Tableau Economique d'ensemble) and the TES (Tableau Entrees
Sorties)
}
\details{
Many arguments enable to make very customized querries in the data
base.
}
\examples{
# by default, the products table
stbl(2023)
# with the nomenclature with 4 sectors for the years 2021, 2022,
# 2023 with only the ressources
stbl(an = 2021:2023, sgn = "pos", naf = "A4")
# the products table for 2020 with the 4 sectors nomenclature and a
# large format
stbl(2023, naf = "A4", format = "large")
# idem, but separate: investment and stock variations, ISBL and
# household, taxes and subventions
stbl(2023, naf = "A4", format = "large",
       net_acc = FALSE, net_tax = FALSE, isbl = TRUE)
# the large format is used with several years
stbl(2021:2023, naf = "A4", format = "large")
# sectors table, two variables (intermediate consumptions and wages
# are selected)
stbl(2020, tableau = "branches", naf = "A4", variables = c("ci.br", "l"))
# production and exploitation accounts with a large format
stbl(2020, tableau = "branches", naf = "A4", format = "large")
# margin format, the production and the intermediate consumptions
# of productive sectors
stbl(2020, tableau = "branches", naf = "A4", format = "marges")
# default TEE: financial and non financial societes are merged in a
# unique agent (`ent`), ISBL are merged with households fusionnees
stbl(2021, tableau = "tee")
# idem with ISBl, financial and non-financial enterprises,
# separation of taxes and subventions
stbl(2021, tableau = "tee", sep_ent = FALSE, isbl = TRUE, net_tax = FALSE)
# selection des menages et de l'Etat, pour les comptes de
# production et d'utilisation du revenu
stbl(2021, tableau = "tee", agents = c("men", "etat"),
       comptes = c("prod", "util"))
# idem, mais selection uniquement des ressources
stbl(2021, tableau = "tee", agents = c("men", "etat"),
       comptes = c("prod", "util"), sgn = "pos")
# a complete TEE with a large format
stbl(2021, tableau = "tee", format = "large")
# property income
stbl(2021, tableau = "tee", variables = "rp")
# net (ressources minus expenses) of property income
stbl(2021, tableau = "tee", variables = "rp", sgn = "diff")
}
