% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wider}
\alias{wider}
\title{Reshape a data frame obtained using \code{stbl}}
\format{
a data frame
}
\usage{
wider(x, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{a data frame obtained using \code{stbl}}

\item{rows}{a character with one or two variables}

\item{cols}{a character with one or two variables}
}
\description{
Up to two variables should be indicated as "time-invariant" and two
variables as "time-varying", the cells are filled with the \code{value}
column
}
