% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtbl.R
\name{xtbl}
\alias{xtbl}
\title{Transformation of the raw table}
\usage{
xtbl(x, formula, shares = NULL, labels = NULL)
}
\arguments{
\item{x}{a data frame,}

\item{formula}{a formula indicating the lines of the resulting
table on the left-hand side and the columns on the right-hand
side}

\item{shares}{a character vector indicating the variables for which
the proportions should be computed}

\item{labels}{a character vector indicating variables for which
labels will be used in the table}
}
\value{
a data frame
}
\description{
The raw table obtained using \code{"stbl"} is in a long format and
contains several characters and one numerical colomn. The raw table
can be nicely formated by pivoting it wider and creating groups of
rows and of columns
}
