% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtbl.R
\name{xtbl2tt}
\alias{xtbl2tt}
\title{A data frame formated as a markdown table using tinytable}
\usage{
xtbl2tt(x, height = 1, ...)
}
\arguments{
\item{x}{a data frame}

\item{height}{la hauteur des lignes}

\item{...}{supplementary arguments passed to \code{tinytable}}
}
\description{
This function is particulary usefull for a table in "wide"
format. It deals with lines and columns groups.
}
