% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sam2tt}
\alias{sam2tt}
\title{SAM formated as a markdown table}
\usage{
sam2tt(
  x,
  rm_zero = TRUE,
  bigmark = " ",
  decimal = ",",
  digits = 1,
  height = 1,
  fontsize = 1,
  colors = list(groupes = "grey70", comptes = "grey85", soldes = "grey95")
)
}
\arguments{
\item{x}{a \code{sam} object}

\item{rm_zero}{a boolean, if \code{TRUE}, zero values are removed from
the matrix}

\item{bigmark}{a character indicating the millions separator}

\item{decimal}{decomal sign (by default the comma)}

\item{digits}{the number of digits}

\item{height}{line height}

\item{fontsize}{police size}

\item{colors}{a named list of colors for the groups, the accounts
and the soldes}
}
\description{
The tinytable package is used and functions of this package can be
used to customize the result. The table can then be coerced in
several format, especially LaTeX and html
}
